gcp/requirements.txt
confluent-kafka==2.1.0
pymongo==4.5.0
requests
fastapi 
uvicorn 
python-dotenv